Intent: Consolidate cache primitives under utils facade Operators involved: [Coherence, Coupling] Affected invariants: [#8, #9] Key changes: - Move CacheManager and cache layer implementations into tnfr.utils.cache and expose a compatibility shim in tnfr.cache. - Update utils exports, stubs, and cache helpers to reference the consolidated module without circular imports. - Refresh documentation and structural cache tests to target tnfr.utils.cache and adjust telemetry logger expectations. Expected risks/dissonances: Potential import regressions if downstreams rely on tnfr.cache; mitigated via compatibility module and updated tests. Metrics: C(t), Si, νf unchanged; cache telemetry remains observable via existing publishers. Equivalence map: tnfr.cache.* → tnfr.utils.cache.* (re-export).
